types: put extended logbooks-related types in logbook.api.types
